+++ bibtextools/__main__.py
@@ -36,6 +36,7 @@ def get_arg_parser():
                         help="If this is set, the title field will be surrounded by curly brackets.")
     parser_modern.add_argument("--iso4", action="store_true",
                         help="If this is set, the journal titles will be abbreviated according to the ISO4 standard.")
+    parser_modern.add_argument("--force", action="store_true", help="Force the automatic removal of duplicate entries (the shorter one will be removed) and skip the interactive prompt.")
     parser_modern.add_argument("-v", "--verbose", action="count", default=0, help="Verbosity level. -v is info and -vv is debug")
     parser_modern.add_argument("-o", "--output", help="Output file for the new bib entries. If not specified, it will be the input file with a 'clean-' prefix.")
     parser_modern.add_argument("bib_file")
@@ -47,6 +48,7 @@ def get_arg_parser():
                         default=DEFAULT_REMOVE,
                         help="Name of fields that should be removed for the clean bib file. By default, this is 'abstract', 'annote', 'file', 'keyword'. Leave empty to not delete any fields")
     parser_clean.add_argument("-u", "--replace_unicode", action="store_true", help='Replace unicode characters by the LaTeX syntax, e.g., ä --> {\\"a}')
+    parser_clean.add_argument("--force", action="store_true", help="Force the automatic removal of duplicate entries (the shorter one will be removed) and skip the interactive prompt.")
     parser_clean.add_argument("-v", "--verbose", action="count", default=0, help="Verbosity level. -v is info and -vv is debug")
     parser_clean.add_argument("-o", "--output", help="Output file for the new bib entries. If not specified, it will be the input file with a 'clean-' prefix.")
     parser_clean.add_argument("bib_file")
@@ -54,7 +56,8 @@ def get_arg_parser():
     parser_combine = subparsers.add_parser("combine", help="Combine multiple bib files into a single one.")
     #parser_combine.add_argument("-a", "--abbr_file", help="Bib-file that contains abbreviations")
     parser_combine.add_argument("-v", "--verbose", action="count", default=0, help="Verbosity level. -v is info and -vv is debug")
-    parser_combine.add_argument("--allow_duplicates", action="store_true", help="If this option is passed, the keys of the bib items remain untouched. This leads to duplicates if the same key is used in multiple files.")
+    parser_combine.add_argument("--allow_duplicates", action="store_true", help="If this option is passed, the IDs of the bib items remain untouched. This leads to duplicate IDs if the same key is used in multiple files. Note that duplicate entries (by content) will still be removed.")
+    parser_combine.add_argument("--force", action="store_true", help="Force the automatic removal of duplicate entries with the same citation (the shorter one will be removed) and skip the interactive prompt.")
     parser_combine.add_argument("-o", "--output", help="Output file for the new bib entries. If not specified, it will be the input file with a 'clean-' prefix.")
     parser_combine.add_argument("bib_files", nargs="+")
     return parser